// frontend/types/stats.ts
export type Stats = {
  day_1_count: number;
  day_2_count: number;
  day_3_count: number;
  day_4_count: number;
  day_5_count: number;
  day_6_count: number;
  day_7_count: number;
  day_8_count: number;
  week_1_count: number;
  week_2_count: number;
  week_3_count: number;
  week_4_count: number;
  week_current_count: number;
  month_1_count: number;
  month_current_count: number;
  all_time_count: number;
};